🛡️ Sentinel: [HIGH] Refine CORS logic to evaluate specific origins first
Updated the CORS origin check so that explicit valid origins in the ALLOWED_ORIGINS list are checked and permitted *before* checking for the presence of a wildcard `*`. If the origin isn't listed but `*` is present, it will log a warning and reject the request (since credentials are true). Also updated the `cors.test.ts` to assert this combined scenario.
